+++ src/views/admin/stage/AdminStageManageEditView.vue
@@ -8,19 +8,15 @@ import { useSnackbarStore } from '@/stores/useSnackbarStore.ts';
 import { useField, useForm } from 'vee-validate';
 import AdminStageService from '@/api/admin/AdminStageService.ts';
 import EditForm from '@/components/form/EditForm.vue';
-import AdminArtistService from '@/api/admin/AdminArtistService.ts';
 import { toTypedSchema } from '@vee-validate/zod';
 import { number, object, string } from 'zod';
+import ArtistSearchDialog, { Artist } from '@/components/dialog/ArtistSelectDialog.vue';
+import ArtistsField from '@/components/form/textfield/ArtistsField.vue';
 
 type UpdateStageForm = {
   startTime: string,
   ticketOpenTime: string,
   artistIds: number[],
-}
-
-type Artist = {
-  id: number,
-  name: string
 }
 
 const route = useRoute();
@@ -50,8 +46,12 @@ onMounted(() => {
 const { isSubmitting, meta, resetForm, setErrors, handleSubmit } = useForm<UpdateStageForm>({
   validationSchema: toTypedSchema(
     object({
-      startTime: string().min(1),
-      ticketOpenTime: string().min(1),
+      startTime: string({
+        required_error: '공연 시작 시간은 필수입니다.',
+      }),
+      ticketOpenTime: string({
+        required_error: '티켓 오픈 시간은 필수입니다.',
+      }),
       artistIds: number().array(),
     }),
   ),
@@ -59,39 +59,7 @@ const { isSubmitting, meta, resetForm, setErrors, handleSubmit } = useForm<Updat
 const startTimeField = useField('startTime');
 const ticketOpenTimeField = useField('ticketOpenTime');
 const artistIdsField = useField<number[]>('artistIds');
-const fetchedArtists = ref<Artist[]>([]);
 const showArtistSelectDialog = ref(false);
-const fetchArtistsLoading = ref(false);
-const artistSearchKeyword = ref('');
-
-function fetchArtists() {
-  AdminArtistService.fetchArtists({
-    page: 1,
-    itemsPerPage: 10,
-    sortBy: [],
-  }, {
-    searchKeyword: artistSearchKeyword.value,
-    searchFilter: 'name',
-  }).then(response => {
-    fetchedArtists.value = response.data.content.map(artist => (
-      { id: artist.id, name: artist.name }
-    ));
-  });
-}
-
-const putArtist = (artist: Artist) => {
-  if (artists.value.has(artist.id)) {
-    snackbarStore.showError('이미 등록된 아티스트 입니다.');
-    return;
-  }
-  artists.value.set(artist.id, artist);
-  artistIdsField.value.value.push(artist.id);
-};
-
-const removeArtist = (artist: Artist) => {
-  artists.value.delete(artist.id);
-  artistIdsField.value.value = [...artists.value.keys()];
-};
 
 const onUpdateSubmit = handleSubmit(async form => {
   try {
@@ -109,7 +77,7 @@ const onUpdateSubmit = handleSubmit(async form => {
   }
 });
 
-function onDeleteSubmit() {
+const onDeleteSubmit = () => {
   AdminStageService.deleteStage(stageId.value!).then(() => {
     snackbarStore.showSuccess('공연이 삭제되었습니다.');
     router.back();
@@ -118,72 +86,16 @@ function onDeleteSubmit() {
       snackbarStore.showError(e.message);
     } else throw e;
   });
-}
+};
 
 </script>
 
 <template>
-  <v-dialog
-    v-model="showArtistSelectDialog"
-    max-width="500"
-  >
-    <v-card class="pa-5">
-      <v-card-title class="text-h5 text-center">
-        아티스트 선택
-      </v-card-title>
-      <v-row :no-gutters="true" align="center">
-        <v-col :cols="10">
-          <v-text-field
-            class="pa-2 ma-2"
-            v-model="artistSearchKeyword"
-            label="아티스트 이름"
-            prepend-inner-icon="mdi-magnify"
-            :single-line="true"
-            variant="outlined"
-            :hide-details="true"
-            maxLength="50"
-          />
-        </v-col>
-        <v-col :cols="2">
-          <v-btn
-            :loading="fetchArtistsLoading"
-            :disabled="!(!!(artistSearchKeyword))"
-            class="py-7 text-h6"
-            color="blue"
-            variant="flat"
-            :block="true"
-            text="검색"
-            @click="fetchArtists"
-          />
-        </v-col>
-      </v-row>
-
-      <v-table class="text-center">
-        <thead>
-        <tr>
-          <th id="name" class="text-center">
-            이름
-          </th>
-          <th id="selectBtn" class="text-center" />
-        </tr>
-        </thead>
-        <tbody>
-        <tr
-          v-for="artist in fetchedArtists"
-          :key="artist.id"
-        >
-          <td>{{ artist.name }}</td>
-          <td>
-            <v-btn
-              @click="putArtist(artist)"
-              text="선택"
-            />
-          </td>
-        </tr>
-        </tbody>
-      </v-table>
-    </v-card>
-  </v-dialog>
+  <ArtistSearchDialog
+    v-model:artists="artists"
+    v-model:show-dialog="showArtistSelectDialog"
+    @append-callback="artist => artistIdsField.value.value.push(artist.id)"
+  />
   <EditForm
     form-title="공연 수정/삭제"
     :loading="isSubmitting"
@@ -191,30 +103,12 @@ function onDeleteSubmit() {
     :on-delete-submit="onDeleteSubmit"
     :is-touched="meta.dirty"
   >
-    <div>
-      <p class="ml-2 ma-1 text-subtitle-2 text-grey-darken-2">아티스트 목록</p>
-      <v-card
-        class="pa-6 mb-6"
-        variant="outlined"
-        @click="showArtistSelectDialog = true"
-      >
-        <v-row>
-          <v-col
-            v-for="artist in artists.values()"
-            :key="artist.id"
-            class="py-1 pe-0"
-            cols="auto"
-          >
-            <v-chip
-              :closable="true"
-              @click:close="removeArtist(artist)"
-            >
-              {{ artist.name }}
-            </v-chip>
-          </v-col>
-        </v-row>
-      </v-card>
-    </div>
+    <ArtistsField
+      :artists="artists"
+      :error-messages="artistIdsField.errorMessage.value"
+      @remove-callback="_ => artistIdsField.value.value = [...artists.keys()]"
+      @click="showArtistSelectDialog = true"
+    />
     <v-text-field
       class="mb-3"
       type="datetime-local"
